Decode gzip or deflate content
Requesting a page from a site that encodes page's content using deflate or gzip will result in a failure to proxy the page successfully (its compressed, not raw text)
Sending the extra header stating you accept no encoding methods (empty string) will *request* that no compression is used in the response. 
Many sites ignore this request header and respond with compressed data anyway (KAT!). 
This attempts to decode the data so it can be proxied.
mh